+++ tests/test_getters.ts
@@ -465,6 +465,47 @@ describe('Jobs getters', function () {
     queue.add('test', { foo: 2 });
   });
 
+  it('should return deduplicated jobs for duplicates types', async function () {
+    queue.add('test', { foo: 1 });
+    const jobs = await queue.getJobs(['wait', 'waiting', 'waiting']);
+
+    expect(jobs).to.be.an('array');
+    expect(jobs).to.have.length(1);
+  });
+
+  it('should return jobs for all types', function (done) {
+    let counter = 0;
+    const worker = new Worker(
+      queueName,
+      async () => {
+        counter++;
+        if (counter == 2) {
+          await queue.add('test', { foo: 3 });
+          return queue.pause();
+        }
+      },
+      { connection },
+    );
+
+    worker.on(
+      'completed',
+      after(2, async function () {
+        try {
+          const jobs = await queue.getJobs();
+          expect(jobs).to.be.an('array');
+          expect(jobs).to.have.length(3);
+          await worker.close();
+          done();
+        } catch (err) {
+          done(err);
+        }
+      }),
+    );
+
+    queue.add('test', { foo: 1 });
+    queue.add('test', { foo: 2 });
+  });
+
   it('should return 0 if queue is empty', async function () {
     const count = await queue.getJobCountByTypes();
     expect(count).to.be.a('number');